[PT] [API] Move TorchLoadOptions to engine (#27467)
Move `TorchLoadOptions` to the `modular.engine` package alongside
`CommonLoadOptions` and `TensorFlowLoadOptions`.
It makes sense to have `TorchLoadOptions` in `modular.engine` now that
we officially support PyTorch.

Note: the only "option" in `TorchLoadOptions` is a list of input specs.
It seems reasonable for ONNX and TF to also support input specs that are
compatible with the graph input specs.
So #27466 tracks moving this option to `CommonLoadOptions`.

As a drive by fix a dependency on engine/__init__.py in the engine CML.

MAX_PYTHON_ORIG_REV_ID: bfbfeb1bbcc20ece64992f6120e3b65c23b99b7f