[EngineAPI] Allow Python to handle Mojo values (#46652)
This is required for Python code to be able to manipulate the KV cache,
which is a MO opaque object. There are currently lifetime issues with
this code, and most uses will currently crash. I have not been able to
debug these yet, but providing a PR to allow others to experiment in the
mean time.

Tests for this code (currently failing due to memory lifetime issues)
are part of a separate PR.

Part of MSDK-843.

MAX_PYTHON_ORIG_REV_ID: d95c4e69e190b33754a5725b4f04383b1b6c55a5