[EngineAPI] Allow Mojo values as input in driver tensor interface (#47195)
Closes MSDK-947.

This rejiggers some of the internal APIs a bit. Before, they were
hard-coded to taking only device tensor memories, and had no way to
accept Mojo values. For the Mojo engine API, there's a dedicated
separate call that allows it to accept Mojo values as well, but it has
some weird restrictions, like that Mojo values must all occur at the end
of the input list, and can't be interleaved before a tensor. I didn't
want this restriction, so I instead switched it to use `M_AsyncValue` as
the non-driver API uses, and then in the bindings I unwrap the driver
tensors to `AnyAsyncValueRef` and re-wrap as `M_AsyncValue` as the
previous API would have done, but now with the flexibility that driver
tensors can be interleaved with any other type of value.

---------

Co-authored-by: Brendan Duke <bduke@modular.com>
MAX_PYTHON_ORIG_REV_ID: 681add28b53907a249d2cc46447dead486f36b28