[Python][API] __call__ API for models (#47776)
This method implements a `__call__` method for models. This method
supports both positional and named arguments with the semantics
described in [this
[Internal link]

We intend to iterate on this method until we reach parity with the
`execute_legacy` method. For now, this should be a reasonable start.

References MSDK-1019

MAX_PYTHON_ORIG_REV_ID: fcf17a314d2047aa7da9c3ab52e3ce8ae2bddeab