[Support][NFC] Change kGPULabel from "cuda" to "gpu" (#51527)
This change modifies how devices specs are defined, changing device
specs for GPUs from using the label `"cuda"` to using the label `"gpu"`.

At the device spec level, we don't care which GPU is used, but we do
care to specify whether a tensor or operation should occur on the CPU or
the GPU. So the use of "cuda" as a label over-specifies the intention.
With the implementation of AMD GPU support (and other GPUs in the
future), we would like to simplify the device spec labels so that we
only specify the necessary information.

This pull request includes many changes, most of them in test code which
are a straightforward find and replace. However, there are changes to
production code necessary as well, although none of them should change
behavior. I'll comment specifically on the product code changes to
highlight them in an attempt to make this PR easier to review.
Unfortunately, I don't see a good way to break this PR up into smaller
chunks to review.

Towards GEX-1269

MAX_PYTHON_ORIG_REV_ID: a6c20f48ded1777e26732455518bc85ce0be4811