[SDK] Add a new private method to use experimental kernels (#55840)
This is used to allow us to turn on / off kernels which are not ready
for production and be able to debug. The immediate use case for this is
the hopper support where the kernels want to be tested, but not ready to
be set to be the default.

See
[Internal link]
for an example of how to use this flag

MAX_PYTHON_ORIG_REV_ID: 5456d2028aad2c672126e447d63f43af45acb700