[SDK] Expose Accelerator.can_access (#57241)
Expose `Device::canAccess` to the Python API.
This is useful in a follow up to stage different ops from the graph API
depending on whether peer access is enabled or not.

An alternative would be to handle this in the
`mo.distributed.allreduce.sum` op lowering.
If that's preferable we can clean it up in a follow up.

MAX_PYTHON_ORIG_REV_ID: fd41fd034318a5cb50c557838c1e4e3f224d0919