[SDK] Unify max._core.dtype.DType and max.dtype.DType (#57521)
Fixes GEX-1848

This starts the monkey-patch approach. We define most of the class in
C++, but for the things that are difficult to move we keep in Python,
and then dynamically add the method to the class.

MAX_PYTHON_ORIG_REV_ID: 1272b3d5966b1efccdfd8e12a7a63b17eece074d