[SDK] Reorder shape and dtype fields of driver Tensor (#58904)
The convention within the codebase is to order fields as `(dtype,
shape)`. The driver Tensor type is inconsistent with that convention.

This change re-orders the dtype and shape fields in the constructor to
be in line with the rest of the codebase.

Closes: [GEX-1987: Standardize argument order for shape and
[Internal link]
MAX_PYTHON_ORIG_REV_ID: bf3cd14d10d1616dbf9f43b57a9f95f2b8db72f8