[python-bindings][SDK] feat: Create `__mojocache__` directory for caching Python/Mojo importer compiled .so (#61031)
This adds a cache to our existing Python import hook for Mojo. Given a
project structure like:

```text
my-project
|-- main.py
\-- mojo_module.mojo
```

When `import mojo_module` is executed within Python, the resulting .so
file will be stored alongside the Mojo source file, in a new
`__mojocache__` directory:

```text
my-project
|-- main.py
|-- mojo_module.mojo
\-- __mojocache__
    \-- mojo_module.hash-91902a236aebcc96.so
```

Then, the next time that `import mojo_module` is evaluated in a
subsequent Python session, the cached `.so` will be loaded instead of
re-compiling the Mojo source from scratch.

The `hash-91902a236aebcc96` portion of the file name contains the
truncated SHA256 hash of the `*.mojo` files (and their names) that were
in the directory that the imported Mojo file was located in. This is a
rough cache key heuristic, but it's hoped will be sufficient to ensure
the Mojo files are re-compiled whenever the user makes a change they
intuitively expect will lead to recompilation.

Future PRs will follow up with refinements to the caching logic, for
example to incorporate the Mojo compiler version in the file name, and
to add logging.

For imported Mojo modules that resolve to an `__init__.mojo` file in a
directory matching the imported module name, the `__mojocache__` will be
created within the Mojo package directory:

```text
my-project
|-- main.py
\-- mojo_module
    |-- __init__.mojo
    \-- __mojocache__
        \-- mojo_module.hash-91902a236aebcc96.so
```

MAX_PYTHON_ORIG_REV_ID: e404dfebd572c1fc5334f805779b7644352dcf0e