[bindings] Update to nanobind 2.7.0 (#61962)
Most of the stubfiles changes are from [this
PR](https://github.com/wjakob/nanobind/pull/938) which stops the sorting
of items in the file.

I'm changing the system to avoid use of the BCR here. While we could
keep using it, we seem to be the only people interested in it, and so it
doesn't get updated. nanobind_bazel seems to be more used, but it's more
work to use that just in order to patch over it. Since our way of
building involves patching over a ton of the BUILD file anyways, let's
just vendor the entire build system.

The extensions have grown slightly, on x64 Linux the 3.12 bindings file
went from 12.60MB to 12.74MB.

MAX_PYTHON_ORIG_REV_ID: 5d1ffda97217ab509814c6cbfabbf598d85b1f0a