[SDK] Remove direct torch dep from max/torch (#66452)
We are planning on enforcing GPU memory for tests. As part of this we
don't know in the test if it needs to dedicate memory to torch or not,
so by default we're giving torch some memory if it is importable. It is
often too importable in tests because of this module, even though they
don't use it.

MAX_PYTHON_ORIG_REV_ID: e12fa5fa38232593d55193dba399a01269c1e982