[Scheduler] Refactor ZMQ to reduce LoC
ZMQ with msgpack serialization requires the concrete types to be passed
in to the ZMQ constructor. Due to limitations of Python 3.9, we also
need
to redundantly specify this type in the generic type annotation.

This has led to a huge propogation of type annotations everywhere in our
codebase because we have not really abstracted this nicely.

I attempt to reduce this proliferation and dump all of the ugly casts
and
such in the `load_scheduler` method. This is terrible but still a step
forward
imo.

Additionally, there was intense duplication of code in each of the ZMQ
types (pull, push, router, dealer). I attempt to factor the
commonalities
into utility methods.

Lastly, I also deleted some features from the ZMQ class including pickle
serialization (yay!), lazy initialization, and timeouts.

I found that lazy initialization was fairly complex and simply a hack to
get around the fact that ZMQ contexts were not serializable. Instead, we
should not even be sending ZMQ sockets across processes in the first
place,
and should send a lightweight ZmqConfig object that contains the basics
needed to construct the ZMQ socket.

I deleted timeouts because it was unused.

MAX_PYTHON_ORIG_REV_ID: 47b682617a972b6117f322073ee4ed71d2ce2157