[Pipelines] Flip switch to disable prefix caching for pixtral
Pixtral models use an IMG tag in the input to the model which is what we
use as our "keys" in the KV Cache. We then expand that image out to the
individual image embeddings inside of the model.

If prefix caching is enabled, we skip this expansion step and just use
the image patches of whatever image was first sent with this prompt.
This change ensures that the default behavior avoids this.

MODULAR_ORIG_COMMIT_REV_ID: 0f6b41c04c05592ae50c49c8301c9c987b5412f8