[SDK] Catch torch exception in dtype.py
Catch all exceptions in dtype.py when attempting to import `torch`.
This is helpful if the `torch` installation is installed but broken in
some way (for example if PYTHONPATH points to max.torch then we'll get
an attribute error on `torch.bool`).

As a drive by, fix an issue in the error message by capturing the
exception as a default argument (previously this would raise a
NameError).

Testing:
- Unit test that covers both fixes.
MAX_PYTHON_ORIG_REV_ID: 603661f1aacc8e37274b2e93738ae947de8c53df