[Stdlib] Generalize the SIMD._simd_apply function
Generalize the simd_apply function by having a default return type
that's the same as the input DType. This is largely as an NFC, but will
allow us to use the function in subsequent PRs in special cases in the
math package.

MODULAR_ORIG_COMMIT_REV_ID: bcecc872d2ad9919a411390197eb93b7fdd1001d