[stdlib] fix leak in os.path.realpath
Previously realpath allocated a char buffer with UnsafePointer before
passing it to realpath, and then sticking it into the guts of a new
string. Unfortunately if realpath returned an error this buffer was
never freed.

We fix this by instead allocating a String of appropriate capacity and
passing its buffer directly into realpath, then setting the length and
nul terminated flag if successful. An internal _ugly function is added
to set this flag.

Note: realpath is somewhat inherently buggy when used with a passed-in
buffer like this. Ideally we'd let realpath allocate the buffer itself.
However, right now mojo doesn't use the same allocator as libc and we
would not be able to use String to free the buffer.
MODULAR_ORIG_COMMIT_REV_ID: 7367cbb4408966617556f19b7860e4a5338607e1