[SDK] Thread per-device chains through control flow
Chains used/updated within MO control-flow operations need to be
reflected in the operands/results of those control-flow ops. If they are
not, it likely means the chain is being dropped erroneously.

With the introduction of per-device chains, the per-device chains are
used to build operations but do not become new results of
mo.if/mo.while. This does not cause correctness issues at the moment
due to `ops.inplace_custom` joining the per-device chains to the global
chain, which is threaded through control-flow ops.

This change updates the construction logic for mo.while and mo.if to
take/produce a value for each live chain maintained by the graph.
Chains which are unused within the control-flow operations are trivially
eliminated by canonicalizers for those operations, resulting in
control-flow ops which take/produce chains only on the devices used
within their body regions.

MODULAR_ORIG_COMMIT_REV_ID: 391f3391af9a2d42d5ed7aa4960f767598ab4470