[Benchmarking] Fix external benchmarking import problem
Part of this is similar to the previous fix, which targeted only
benchmark_serving but not benchmark_serving_lib, but a naive application
of that does not fully fix the problem. An import needed to be changed
to relative, and usages across the code base needed to be made
consistent as well.

MODULAR_ORIG_COMMIT_REV_ID: f328c51fd118cf4a3cfbf88791a08be668905972