[Benchmarking] Consistently Clip MAX K in Top K Kernels
To ensure that we consolidate the logic surrounding clipping for valid
ranges of top k to the kernel, we passed a previous change which allowed
passing -1 values of top_k to the top_k kernels. Each kernel would then
be responsible for handling the -1 appropriately for the bounds of its
specific implementation. However, this was applied inconsistently
leading to a few OOB with specific request parameter combinations.

This PR consistently clips max_k and top_k values across the Top K
kernels to fix these concerns. It also makes a small change to the
benchmarking payload to ensure None values are not passed in a request
to the server.

MODULAR_ORIG_COMMIT_REV_ID: 921b7ce9bbbc3d4ade9161d83b9f6c4842223e0b