[SDK] Place RoPE on CPU for Gemma 3
Place RoPE on CPU for Gemma 3.
Otherwise a device-to-device is staged in each layer and materialized in
the model execute block.

Note that you may ask why the DtoD transfer isn't hoisted to init.
It is because we currently chain DtoDs with collective operations to
prevent correctness issues (deadlocks).
The chaining and the order of op staging here, prevents the hoisting.
This can be improved in future work.
See:
[Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 711f3cfd90ba6b6ff5a1b9453b256cc624703a45