[Sampling] Move Bitmask Unpacking to Init
We never update the bitmask once provided to FusedSamplingProcessor, and
yet unpack it on every iteration. While it is not causing a problem,
this has the potential to lead to a bug, so I am cleaning it up early,
unpacking the tensor once, and raising if a bitmask is passed to the
sampler in a multi-step scenario.

MODULAR_ORIG_COMMIT_REV_ID: ca4251d57e407fca37a91f20817f881c0819f31f