[Scheduler][DP] Handle chunked requests correctly
Previously, we only checked whether the last request in the batch was
chunked to decide if it should be returned to the prefill queue. With
DP, we now need to check whether the last request of each device’s local
batch is chunked.

MODULAR_ORIG_COMMIT_REV_ID: 5486207ba20c0737248c5bdff8966d3600038417