[bazel] Allow cross platform dynamic mode for cc actions
When using dynamic execution to improve tight loop development
performance, if the remote executors have a different OS / arch from the
host, you will get the wrong binaries for the action.

https://github.com/bazelbuild/bazel/issues/25221

With this change some of the CC tools will work with any host by
providing all the compilers and execing based on the current host. This
approach only works if these are the only tools used in the action.

For example if you're on a macOS host and you're targeting a remote
Linux GPU machine, with remote Linux executors, some compilations can
now also be done locally. Technically this might result in different bit
for
bit outputs, but we trust the hermiticity of our toolchain enough to
risk that
for the improvement in perf.

MODULAR_ORIG_COMMIT_REV_ID: 19fe5d416f767a91e1a8b3bfeb917f0768019f0c