[Kbench][Bug]Fix for crash when attempting to read csv
Flaky failures encountered causing crashes when trying to read the
`output.csv` of a specific benchmark binary being run.
This was reproduced consistently by running with `kbench` with timeout
of 1 second and adding a 2 second sleep before the write in
`open-source/max/mojo/stdlib/stdlib/benchmark/bencher.mojo:1003`.

Since the timeout is passed to the `subprocess.run()` which launches the
benchmarks and creates the csv it is possible that the process might be
killed AFTER the csv is created but BEFORE the data is flushed and
written to disk.

Current version of `get_valid_specs` assumes that if the file exists the
csv can be read. Proposing adding a check for the return code before
reading to account for the described scenario.

MODULAR_ORIG_COMMIT_REV_ID: ce072495d5e4f655d51b1bb5fc1368ab1d836f3c