[MOGG][Models] Preserve mo.rebind when promoting to static dimensions
This fix addresses a bug in the translation of MO to MOGG which removes
mo.rebind ops when they are require for promoting from a symbolic
dimension to a
static dimension. This is required for the mo.rebind that is introduced
when a
rmo.mo.slice is lowered to a mo.slice + mo.rebind.

* Updates the MOToMOGGPass implementation to contain an additional
condition
for keeping mo.rebind ops, specifically when the input type has a
symbolic
dimension and the correspoonding dimension in the output type is static.
* Updates the existing rebind lit test
Conversion/MOToPrimitives/rebind.mlir to
account for this.
* Reverts a work-around in mo.quantize_dynamic_scaled_float8 to avoid
this
issue.

MODULAR_ORIG_COMMIT_REV_ID: a188b623564f62d70253cd231bd6a4c593ca537b