[Models] Delete MTP weights when loading DeepSeekV3 checkpoint
The MTP layer is used for speculative decoding, but we don't support
this yet in the model. So for now, delete the weights from this layer in
order to use strict=True when weight loading. The DeepSeek's inference
script also deletes the weights from this layer in the converter:
https://github.com/deepseek-ai/DeepSeek-V3/blob/4592be48c07f036b32ef971474068aebc489e3e7/inference/convert.py#L53-L54

MODULAR_ORIG_COMMIT_REV_ID: ae4c8108aadf880e6c3b876bea5cd2952472ed6e